requirements.txt
beautifulsoup4==4.14.3
ebooklib==0.20
faiss_gpu==1.7.2
fastapi==0.133.0
fitz==0.0.1.dev2
generator==0.1-alpha
httpx==0.28.1
numpy==2.4.2
pandas==3.0.1
pdfplumber==0.11.9
Pillow==12.1.1
pydantic==2.12.5
pytesseract==0.3.13
python-dotenv==1.2.1
python_docx==1.2.0
Requests==2.32.5
sentence_transformers==5.2.2
torch==2.5.1+cu121
torch==2.10.0
trafilatura==2.0.0
